AdminPage wip3. Functionality for admin to create, update, delete toilets done.
For Update/Delete toilet, admin has to first get the toilet info by ID. If valid, the form then auto populates the input fields with details of the toilet. The admin can then modify the toilet details to update, or delete this toilet. The delete toilet button when clicked, renders
a confirmation button.