Navbar wip4. Created logout functionality and conditional renderings.
The login and register buttons render when user is not logged in.
If user is logged in, the login and register buttons disappears and is replaced by a logout button instead.
The logout button calls sends user back to home and also calls a function that will clear localStorage and sets isLoggedIn to false.